standardize MJD warnings so that they can be filtered
Previously, when we attempted to transform UTC to UT1 on a UTC
that was outside the bounds of astropy.time's interpolation,
a customized warning was raised which listed the offending UTC
value.  This made it impossible to filter the warnings, since
each individual warning had a different message and thus appeared
to python as unique warning.  With this commit, the warning message
is made uniform so that simple warning filters can catch it.