simplify PhoSimMetaData in ObservationMetaData
Store extra metadata needed by PhoSim in a plain dict.  Do not
use that dict to intialize the other data in the ObservationMetaData.
Add a method write_phoSim_metadata() to CompoundCoordinateTransformations.py.
This method takes an ObservationMetaData and a file handle and writes the
header expected by PhoSim to the file_handle.  This will be used by the
PhoSim InstanceCatalog classes.