remove epoch from ObservationMetaData
We were not using it.  I think the idea was, if we were ever
connecting to a database that stored RA, Dec in some system other
than J2000, we would need to transform from one system to another.
That use case hasn't come up yet.  If it ever does, we can put epoch
back in (assuming that is the best way to deal with that case).  For
now, I don't want to maintain a feature that isn't being used (or
doesn't even really exist as a "feature").